+++ src/scope/migrations/component-version/index.js
@@ -1,4 +1,5 @@
 import addBindingPrefixDeclartaion from './add-binding-prefix';
 import changeVersionToSemVerDeclartaion from './to-semver-migration';
+import changeImportSpecifiersToArray from './change-import-specifiers-to-array';
 
-export { addBindingPrefixDeclartaion, changeVersionToSemVerDeclartaion };
+export { addBindingPrefixDeclartaion, changeVersionToSemVerDeclartaion, changeImportSpecifiersToArray };