+++ src/cli/commands/public-cmds/commit-cmd.js
@@ -2,6 +2,7 @@
 import Command from '../../command';
 import { commitAction, commitAllAction } from '../../../api/consumer';
 import Component from '../../../consumer/component';
+import ModelComponent from '../../../scope/models/component';
 
 const chalk = require('chalk');
 
@@ -46,13 +47,16 @@ export default class Export extends Command {
     return commitAction({ id, message, force, verbose, ignoreMissingDependencies: ignore_missing_dependencies });
   }
 
-  report(components: Component | Component[]): string {
-    if (!components) return chalk.yellow('nothing to tag');
-    if (!Array.isArray(components)) components = [components];
-
+  report(results): string {
+    if (!results) return chalk.yellow('nothing to tag');
+    const {
+      components,
+      autoUpdatedComponents
+    }: { components: Component[], autoUpdatedComponents: ModelComponent[] } = results;
     function joinComponents(comps) {
       return comps
         .map((comp) => {
+          if (comp instanceof ModelComponent) return comp.id();
           // Replace the @1 only if it ends with @1 to prevent id between 10-19 to shown wrong ->
           // myId@10 will be myId0 which is wrong
           return comp.id.toString().endsWith('@1') ? comp.id.toString().replace('@1', '') : comp.id.toString();
@@ -74,12 +78,20 @@ export default class Export extends Command {
 
     const changedComponents = components.filter(component => component.version > 1);
     const addedComponents = components.filter(component => component.version === 1);
+    const autoUpdatedCount = autoUpdatedComponents ? autoUpdatedComponents.length : 0;
 
     return (
-      chalk.green(`${components.length} components tagged`) +
-      chalk.gray(` | ${addedComponents.length} added, ${changedComponents.length} changed\n`) +
+      chalk.green(`${components.length + autoUpdatedCount} components tagged`) +
+      chalk.gray(
+        ` | ${addedComponents.length} added, ${changedComponents.length} changed, ${autoUpdatedCount} auto-tagged\n`
+      ) +
       outputIfExists(addedComponents, 'added components: ') +
-      outputIfExists(changedComponents, 'changed components: ', true)
+      outputIfExists(changedComponents, 'changed components: ', true) +
+      outputIfExists(
+        autoUpdatedComponents,
+        'auto-tagged components (as a result of tagging their dependencies): ',
+        true
+      )
     );
   }
 }