+++ src/prompts/schemas/user-password.js
@@ -1,0 +1,16 @@
+/** @flow */
+
+/**
+ * schema for passphrase prompt on SSH.
+ */
+export default {
+  properties: {
+    username: {
+      required: true
+    },
+    password: {
+      hidden: true,
+      required: true
+    }
+  }
+};