+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- bug fix - generated .ts links were not valid
+
 ## [0.11.0-test.3] - 2017-11-06
 
 - change versions numbers to be semantic versions
-- export / import performence and stability improvements
+- export / import performance and stability improvements
 - fix a bug which prevent the ci running tests in some cases
 - add plugin mechanism to support different file types
 - bug fix - re-adding a component after exporting it was considered as a new component