+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- bug fix - dependencies were not written to the file-system when cloning a project with an existing bit.map file
 - disable the local search
 - add plugin mechanism to support different file types
 - bug fix - re-adding a component after exporting it was considered as a new component